LayoutTests/editing/pasteboard/drag-and-drop-color-input-events-expected.txt

This test verifies that the input and change events fire when changing the value of a color input via drag and drop. To test manually, drag the color input on the left onto the color input on the right.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



PASS src.value is "#ff0000"
PASS dest.value is "#ff0000"
PASS successfullyParsed is true

TEST COMPLETE
